fix: fflorent review adjustements
